Rename mix lint to mix check_code
